Fix wan_2_2_remix_gguf: cast remaining meta tensors to compute_dtype on init
Meta tensors not in the GGUF file were initialized as float32 by default,
causing dtype mismatches with bfloat16 inputs. Cast them to compute_dtype.

+++ wan_2_2_remix_gguf/pytorch/loader.py
@@ -157,14 +157,16 @@ class ModelLoader(ForgeModel):
                         if param is not None and param.is_meta:
                             module._parameters[pname] = torch.nn.Parameter(
                                 torch.empty(
-                                    param.shape, dtype=param.dtype, device=device
+                                    param.shape,
+                                    dtype=compute_dtype,
+                                    device=device,
                                 ),
                                 requires_grad=param.requires_grad,
                             )
                     for bname, buf in list(module._buffers.items()):
                         if buf is not None and buf.is_meta:
                             module._buffers[bname] = torch.empty(
-                                buf.shape, dtype=buf.dtype, device=device
+                                buf.shape, dtype=compute_dtype, device=device
                             )
             return _orig_dispatch(model, device_map=device_map, **kwargs)
